+++ frontend/components/Logo/index.jsx
@@ -4,15 +4,15 @@ import { withRoute } from '@shopgate/engage/core';
 import { appConfig } from '@shopgate/engage';
 import styles from './style';
 import connect from './connector';
-import { showSwitcherInHeader } from '../../config';
 
 /**
  * The Logo component renders the logo if the switch is visible.
+ * @param {Object} props The component props.
  * @param {boolean} props.isSwitchVisible Indicates whether the switch is visible or not.
  * @returns {JSX.Element}
  */
 const Logo = ({ isSwitchVisible, children }) => (
-  showSwitcherInHeader && isSwitchVisible ? (
+  isSwitchVisible ? (
     <div className={`${styles.container} engage__logo`}>
       <img
         className={styles.image}